# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Machina API",
    "version": "1.0.0",
    "description": "REST API for Linux hypervisor hosts: QEMU/KVM domains via libvirt, networks, storage, host networking, backups, metrics, and optional KubeVirt bundle/apply workflows when configured."
  },
  "servers": [
    { "url": "/api/v1", "description": "Default API base" }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "machina_session",
        "description": "Cookie-based session authentication (`machina_session=`)"
      }
    },
...
  },
  "paths": {
...
    "/backups": {
      "get": {
        "summary": "List all backups",
        "tags": ["Backups"],
        "responses": { "200": { "description": "Backup list" } }
      },
      "post": {
        "summary": "Trigger a backup",
        "tags": ["Backups"],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "vm_name": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Backup triggered" } }
      }
    },
    "/backups/restore": {
      "post": {
        "summary": "Restore from a backup",
        "tags": ["Backups"],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "backup_id": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Restore initiated" } }
      }
    },
...
    "/backups/{id}": {
      "delete": {
        "summary": "Delete a backup",
        "tags": ["Backups"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Backup deleted" } }
      }
    },
    "/backups/{id}/status": {
      "get": {
        "summary": "Get backup status",
        "tags": ["Backups"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Backup status" } }
      }
    },
    "/backups/{id}/verify": {
      "post": {
        "summary": "Verify a backup",
        "tags": ["Backups"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Verification result" } }
      }
    },
    "/backups/{id}/download": {
      "get": {
        "summary": "Download a backup archive",
        "tags": ["Backups"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Backup file stream" } }
      }
    },
...
    "/alerts/{id}/ack": {
      "post": {
        "summary": "Acknowledge an alert",
        "tags": ["Automation"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Alert acknowledged" } }
      }
    },
...
  }
}